+++ module/helpers/metahelpers.mjs
@@ -79,6 +79,7 @@ export function metaLogDocument(app, buttons) {
 }
 
 /**
+ *
  * Helper function to extract the number of d10 dice from a given value
  *
  * @param {String} value
@@ -122,8 +123,10 @@ export async function metaIsItemEquipped(actor, itemName) {
 }
 
 /**
- * ! confirm usage
+ *
  * Helper function to check if a Metapower with a given name is equipped by an actor
+ * Returns true/false
+ *
  * @param {*} actor - Object of the actor
  * @param {*} metapower - String of the metapower name
  * @returns true/false
@@ -160,65 +163,11 @@ export async function metaGetMaxMetapowerLevel(actor, metapower) {
  * @param {*} formElement
  * @returns
  */
- export function metaExtractFormData(formElement) {
+export function metaExtractFormData(formElement) {
 	const formData = new FormData(formElement);
 	let extractedData = {};
 	for (let [key, value] of formData.entries()) {
 		extractedData[key] = value;
 	}
 	return extractedData;
-}
-
-/**
- *! unused - not working yet
- * Helper function to help in resizing elements based on the window size for a Responsive UI
- * To be used in Actor Sheets to start, will extend and addapt it down the line for Item & Effect Sheets
- *
- * @param {*} event
- * @param {*} element
- *
- */
-export function metaHandleResize(event, element, currentWidth) {
-	const windowSize = determineWindowSize(currentWidth);
-	let uiTargets = element ? $(element).find(".meta-ui-responsive") : this.element.find(".meta-ui-responsive");
-	function determineWindowSize(currentWidth) {
-		if (currentWidth <= 280) return "singleColumn";
-		if (currentWidth > 280 && currentWidth <= 450) return "small";
-		if (currentWidth > 450 && currentWidth <= 750) return "medium";
-		if (currentWidth > 750 && currentWidth <= 1190) return "default";
-		if (currentWidth > 1190) return "extended";
-	}
-	function responsiveUIChanges(target, windowSize) {
-		const classToggles = {
-			singleColumn: {
-				"meta-ui-hidden": ["style-cs-rolls", "meta-ui-extra-fields"],
-				"meta-ui-single-column-font": ["style-cs-chars-label", "style-cs-stats-label"],
-				"meta-ui-small-font": false,
-				"meta-ui-medium-font": false,
-				"layout-container-outer-charstats-single-column": "layout-container-outer-charstats",
-				"style-cs-rolls-single-column": "style-cs-rolls",
-			},
-			small: {
-				"meta-ui-hidden": ["meta-ui-hide-at-minimum"],
-				"meta-ui-small-font": ["style-cs-chars-label", "style-cs-stats-label"],
-				"meta-ui-medium-font": false,
-				"meta-ui-single-column-font": false,
-				"layout-container-outer-charstats-minimum": "layout-container-outer-charstats",
-			},
-			// Define other size categories following the same pattern...
-		};
-		// Apply the toggles based on the current window size category
-		const toggles = classToggles[windowSize];
-		for (const [toggleClass, affectedClasses] of Object.entries(toggles)) {
-			if (Array.isArray(affectedClasses)) {
-				affectedClasses.forEach((cls) => target.classList.toggle(cls, toggleClass === true));
-			} else {
-				target.classList.toggle(toggleClass, affectedClasses);
-			}
-		}
-	}
-	// Loop through the uiTargets and apply the appropriate class changes
-	for (let target of uiTargets) {
-		responsiveUIChanges(target, windowSize);
-	}
 }